ai_resume_portal/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
Pillow==10.1.0
PyPDF2==3.0.1
pdfminer.six==20221105
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
nltk==3.8.1

# Environment Variables
python-dotenv==1.0.0

# Database
dj-database-url==2.1.0

# Static Files (for Render)
whitenoise==6.6.0

# Production Server
gunicorn==21.2.0

# Security
django-cors-headers==4.3.1
